Safari-611.1.10/LayoutTests/fast/canvas/webgl/webgl2-context-creation-expected.txt

Checks that using a WebGL 2 context doesn't cause an error.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS context.getError() is context.NO_ERROR
PASS successfullyParsed is true

TEST COMPLETE
